requirements.txt
flask==3.0.0
pandas==2.1.4
xlsxwriter==3.1.9
gunicorn==21.2.0
openpyxl==3.1.2
ftfy==6.1.1
fpdf2==2.7.7